+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * concerto.saml.getAuthenticatedUser fix for serialized session runner service
 * reduced memory consumption on collection fetches
 * item bank extra fields can now be set for flat item bank table
+* removed automatic data connections
 
 #### Improvements